Address access manager review invariants
