+++ docs/domains/access-and-accounts/architecture/data_model.md
@@ -51,7 +51,7 @@ approvals:
 
 Инварианты:
 - в установке должна быть ровно одна активная организация-владелец;
-- организацию-владельца нельзя архивировать, отключать, удалять или переводить в `suspended`;
+- организация-владелец создаётся и остаётся только в статусе `active`;
 - клиентские организации и организации внешних исполнителей не получают прав на платформенный контур без явных правил;
 - архивирование организации публикует событие для других сервисов.
 
@@ -95,6 +95,11 @@ approvals:
 | `status` | enum | no | `active`, `disabled`. |
 | `version` | int64 | no | Конкурентные изменения. |
 
+Инварианты:
+- `disabled` запись allowlist является явным отказом для совпавшего email или домена;
+- точное совпадение по email имеет приоритет над доменным совпадением;
+- по `disabled` записи пользователь не создаётся даже в статусе `pending`.
+
 ### `Group`
 
 | Поле | Тип | Может быть пустым | Примечание |
@@ -111,6 +116,7 @@ approvals:
 
 Инварианты:
 - родительская группа должна находиться в той же области, что и дочерняя;
+- `slug` уникален внутри области: отдельно для `global` и отдельно внутри каждой организации;
 - иерархия групп не должна содержать циклов.
 
 ### `Membership`
@@ -153,12 +159,15 @@ approvals:
 | `resource_type` | string | no | Тип ресурса: `project`, `repository`, `package`, `runtime`, другое. |
 | `resource_id` | string | yes | Внешний идентификатор ресурса. |
 | `scope_type` | enum | no | `global`, `organization`, `project`, `repository`. |
-| `scope_id` | string | yes | Внешний идентификатор области. |
+| `scope_id` | string | yes | Внешний идентификатор области; для `global` всегда пустой. |
 | `priority` | int | no | Для явных исключений. |
 | `status` | enum | no | `active`, `disabled`. |
 | `version` | int64 | no | Конкурентные изменения. |
 
-Инвариант: явный запрет побеждает разрешение, если оба правила применимы к одному действию и ресурсу.
+Инварианты:
+- явный запрет побеждает разрешение, если оба правила применимы к одному действию и ресурсу;
+- правило с `scope_type=global` и пустым `scope_id` применяется ко всем областям;
+- правило можно создать только для активного `AccessAction`.
 
 ### `ExternalProvider`
