+++ docs/domains/access-and-accounts/delivery/wave7_access_and_accounts.md
@@ -67,7 +67,8 @@ approvals:
 | Область | Статус | Где закрывается |
 |---|---|---|
 | gRPC `AccessManagerService`, OpenAPI и AsyncAPI на весь согласованный объём | Готово как стабильный контракт `v1`. | API-контракт домена и машинно-проверяемые спецификации. |
-| Сервисный каркас `access-manager`, конфигурация, health/readyz/metrics и пустой gRPC-сервер | Готово как базовый срез. | Текущий срез реализации. |
+| Сервисный каркас `access-manager`, конфигурация, health/readyz/metrics и пустой gRPC-сервер | Готово как базовый процесс без зарегистрированных RPC-обработчиков. | Текущий срез реализации. |
+| gRPC/HTTP наблюдаемость, трассировка, recovery и единая граница обработки ошибок | В бэклоге; добавляется до регистрации первых RPC/HTTP-обработчиков. | #599, #600, #601, #602 до подключения транспорта. |
 | Миграция БД для основных таблиц домена и outbox | Готово как начальная схема. | #599, #600, #601, #602. |
 | Доменный слой организаций, групп, членства, allowlist, внешних аккаунтов, правил доступа, `CheckAccess` и `ResolveExternalAccountUsage` | Реализация начата; инварианты и модульные тесты покрывают базовый путь. | #599, #600, #601, #602. |
 | PostgreSQL-репозиторий для всех методов доменного интерфейса | В бэклоге: текущий SQL-срез покрывает только часть чтений и записей. | #599, #600, #601, #602 до подключения транспорта. |
