+++ services/internal/access-manager/internal/domain/types/query/query.go
@@ -1,3 +1,4 @@
+// Package query contains typed repository filters for access-manager reads.
 package query
 
 import (
@@ -7,11 +8,13 @@ import (
 	"github.com/codex-k8s/kodex/services/internal/access-manager/internal/domain/types/value"
 )
 
+// MembershipGraphFilter selects active memberships for subject graph expansion.
 type MembershipGraphFilter struct {
 	Subject value.SubjectRef
 	Status  enum.MembershipStatus
 }
 
+// MembershipIdentity is the natural key of a membership edge.
 type MembershipIdentity struct {
 	SubjectType enum.MembershipSubjectType
 	SubjectID   uuid.UUID
@@ -19,6 +22,7 @@ type MembershipIdentity struct {
 	TargetID    uuid.UUID
 }
 
+// AccessRuleFilter selects policy rules applicable to an access check.
 type AccessRuleFilter struct {
 	Subjects     []value.SubjectRef
 	ActionKey    string
@@ -27,6 +31,7 @@ type AccessRuleFilter struct {
 	Scope        value.ScopeRef
 }
 
+// ExternalAccountUsageFilter selects an account binding for a requested usage.
 type ExternalAccountUsageFilter struct {
 	ExternalAccountID uuid.UUID
 	ActionKey         string
