+++ services/internal/access-manager/internal/repository/postgres/access/repository.go
@@ -1,3 +1,4 @@
+// Package access embeds SQL query files for the PostgreSQL access repository.
 package access
 
 import "embed"
