+++ services/internal/access-manager/internal/repository/postgres/access/sql/access.sql
@@ -42,7 +42,7 @@ WHERE (subject_type, subject_id) IN (SELECT subject_type, subject_id FROM candid
   AND action_key = @action_key
   AND resource_type = @resource_type
   AND (resource_id = '' OR resource_id = @resource_id)
-  AND (scope_type = '' OR (scope_type = @scope_type AND scope_id = @scope_id))
+  AND ((scope_type = 'global' AND scope_id = '') OR (scope_type = @scope_type AND scope_id = @scope_id))
   AND status = 'active'
 ORDER BY priority DESC, updated_at DESC;
